requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest==7.3.1
Faker==37.0.0
requests==2.32.3
